Andreas Putzo <andreas@putzo.net>
in dem Debian Paket ab Version 2.09 gab es einen Patch, der für
dl_mysql_real_connect den Parameter unsigned int clientflag auf unsigned long
clientflag setzt.
Das scheint auch richtig zu sein, zumindest nach 
http://dev.mysql.com/doc/refman/5.0/en/mysql-real-connect.html




git-svn-id: https://gpsdrive.svn.sourceforge.net/svnroot/gpsdrive/trunk@1739 e3ca9d6f-f84a-410d-b97a-e7ddc331358d

+++ src/gpsdrive.h
@@ -247,7 +247,7 @@ MYSQL * (*dl_mysql_real_connect)(MYSQL *mysql, const char *host,
                                            const char *db,
                                            unsigned int port,
                                            const char *unix_socket,
-                                           unsigned int clientflag);
+                                           unsigned long clientflag);
 void    (*dl_mysql_close)(MYSQL *sock);
 int     (*dl_mysql_query)(MYSQL *mysql, const char *q);
 my_ulonglong (*dl_mysql_affected_rows)(MYSQL *mysql);
